requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
Pillow
django-extensions
python-decouple
celery
redis
django-channels
channels-redis
cryptography
reportlab
openpyxl
requests
stripe
whitenoise
schedule
